rename modal actions
