+++ frontend/components/auth/AuthContainer.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
-import { logout } from "../../actions/session_actions"
-import { openModal } from "../../actions/modal_actions";
+import { logout } from "../../actions/session_actions";
+import { openModal } from "../../actions/modalActions";
 import Auth from "./Auth";
 
 const mapStateToProps = (state) => {
@@ -8,15 +8,15 @@ const mapStateToProps = (state) => {
   let allUsers = state.entities.users;
   //AuthContainer passes as props to the presentational component currentUser from the state
   return {
-    currentUser: allUsers[currentUserId]
-  }
+    currentUser: allUsers[currentUserId],
+  };
 };
 
 const mapDispatchToProps = (dispatch) => {
   return {
     logout: () => dispatch(logout()),
-    openModal: (modal) => dispatch(openModal(modal))
-  }
+    openModal: (modal) => dispatch(openModal(modal)),
+  };
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Auth);
