+++ frontend/reducers/modal_reducer.js
@@ -1,11 +1,9 @@
-import { OPEN_MODAL, CLOSE_MODAL } from "../actions/modal_actions";
+import { OPEN_MODAL, CLOSE_MODAL } from "../actions/modalActions";
 
-
-
-                                    //modal starts closed
+//modal starts closed
 const ModalReducer = (state = null, action) => {
   Object.freeze(state);
-  switch(action.type) {
+  switch (action.type) {
     case OPEN_MODAL:
       return action.modal;
     case CLOSE_MODAL:
@@ -14,7 +12,7 @@ const ModalReducer = (state = null, action) => {
     default:
       //modal is left intact not opened
       return state;
-  };
+  }
 };
 
 export default ModalReducer;
